--- a/.github/workflows/main.yml
+++ b/.github/workflows/main.yml
@@ -17,7 +17,7 @@ jobs:
     strategy:
       matrix:
         os: [ ubuntu-latest, macos-latest, windows-latest ]
-        go: [ '1.15', 'stable', 'oldstable' ]
+        go: [ '1.20.x', '1.21.x', '1.22.x' ]
 
     runs-on: ${{ matrix.os }}
     steps:
@@ -42,17 +42,7 @@ jobs:
       run: |
         go mod verify
 
-    - name: Test Go 1.15
-      # This step if needed because -shuffle not available on Go 1.15.
-      # Tests are failing on MacOS. So, we just disable it.
-      if: >-
-        matrix.go == '1.15' && matrix.os != 'macos-latest'
-      run: |
-        go test -v -race -cover ./...
-
     - name: Test
-      if: >-
-        matrix.go != '1.15'
       run: |
         go test -v -race -shuffle=on -cover ./...