--- a/cipher.go
+++ b/cipher.go
@@ -25,7 +25,7 @@ func Cipher(payload []byte, mask [4]byte, offset int) {
 	// Count number of bytes will processed one by one from the beginning of payload.
 	ln := remain[mpos]
 	// Count number of bytes will processed one by one from the end of payload.
-	// This is done to process payload by 8 bytes in each iteration of main loop.
+	// This is done to process payload by 16 bytes in each iteration of main loop.
 	rn := (n - ln) % 16
 
 	for i := 0; i < ln; i++ {